Moving enums to common project
