+++ api/src/main/kotlin/com/example/todo/domain/entity/TaskEntity.kt
@@ -1,6 +1,6 @@
 package com.example.todo.domain.entity
 
-import com.example.todo.domain.enumration.TaskStatus
+import com.example.todo.enums.TaskStatus
 import java.time.ZonedDateTime
 
 data class TaskEntity(
