+++ common/src/main/kotlin/com/example/todo/enums/UserRole.kt
@@ -1,4 +1,4 @@
-package com.example.todo.domain.enumration
+package com.example.todo.enums
 
 enum class UserRole {
     ADMIN, USER, INVALID;
